fix(codex): resolve relay CI follow-ups
