+++ extensions/codex/src/app-server/event-projector-tool-transcript.ts
@@ -1,26 +1,23 @@
 import path from "node:path";
 import {
   embeddedAgentLog,
-  runAgentHarnessAfterToolCallHook,
   type AgentMessage,
   type EmbeddedRunAttemptParamsV2 as EmbeddedRunAttemptParams,
 } from "openclaw/plugin-sdk/agent-harness-runtime";
 import type { Usage } from "openclaw/plugin-sdk/llm";
-import { asDateTimestampMs } from "openclaw/plugin-sdk/number-runtime";
+import { CodexAfterToolCallProjection } from "./event-projector-after-tool-call.js";
 import {
   isMutatingNativeToolItem,
   isNonSuccessItemStatus,
   itemName,
   itemStatus,
   shouldRecordNativeToolTranscript,
-  shouldSynthesizeToolProgressForItem,
 } from "./event-projector-items.js";
+import type { CodexAppServerEventProjectorOptions } from "./event-projector-options.js";
 import {
-  isNativePostToolUseRelayItem,
   itemMeta,
   itemOutputText,
   itemToolArgs,
-  itemToolError,
   itemToolResult,
   itemTranscriptResultText,
 } from "./event-projector-tool-items.js";
@@ -44,7 +41,6 @@ import {
 } from "./protocol.js";
 import { readCodexMirroredSessionHistoryMessages } from "./session-history.js";
 import { sanitizeCodexToolArguments } from "./tool-progress-normalization.js";
-import type { CodexTrajectoryRecorder } from "./trajectory.js";
 import type { CodexTranscriptCheckpointEntry } from "./transcript-checkpoint.js";
 import { attachCodexMirrorIdentity } from "./upstream-prompt-provenance.js";
 
@@ -145,8 +141,7 @@ export class CodexToolTranscriptProjection {
   private readonly trajectoryResultIds = new Set<string>();
   private readonly trajectoryNamesById = new Map<string, string>();
   private readonly trajectoryItemsById = new Map<string, CodexThreadItem>();
-  private readonly afterToolCallObservedItemIds = new Set<string>();
-  private readonly pendingFileChangeAfterToolCallObservations = new Map<string, CodexThreadItem>();
+  private readonly afterToolCall: CodexAfterToolCallProjection;
   private readonly nativeMcpAppResultDetails = new Map<string, unknown>();
   private readonly nativeMcpAppResultDetailsAttempted = new Set<string>();
   private readonly approvalReviewsByCallId = new Map<string, ToolApprovalReviewState>();
@@ -160,16 +155,12 @@ export class CodexToolTranscriptProjection {
     private readonly turnId: string,
     private readonly progress: CodexToolProgressProjection,
     private readonly nextTranscriptTimestamp: () => number,
-    private readonly options: {
-      nativePostToolUseRelayEnabled?: boolean;
-      resolveNativeFileChangeAfterToolCallCoverage?: (
-        toolUseId: string,
-      ) => "native_apply_patch" | "intercepted" | "pending";
-      prepareNativeMcpAppResultDetails?: (item: CodexThreadItem) => Promise<unknown>;
-      trajectoryRecorder?: CodexTrajectoryRecorder | null;
+    private readonly options: CodexAppServerEventProjectorOptions & {
       checkpointMessage?: (entry: CodexTranscriptCheckpointEntry) => void;
     } = {},
-  ) {}
+  ) {
+    this.afterToolCall = new CodexAfterToolCallProjection(params, progress, options);
+  }
 
   get transcriptMessages(): readonly AgentMessage[] {
     return this.messages;
@@ -503,83 +494,11 @@ export class CodexToolTranscriptProjection {
   }
 
   emitAfterToolCallObservation(item: CodexThreadItem): void {
-    if (
-      item.type === "fileChange" &&
-      this.options.nativePostToolUseRelayEnabled &&
-      !this.afterToolCallObservedItemIds.has(item.id)
-    ) {
-      const coverage =
-        this.options.resolveNativeFileChangeAfterToolCallCoverage?.(item.id) ?? "pending";
-
-      if (coverage === "native_apply_patch") {
-        this.afterToolCallObservedItemIds.add(item.id);
-        return;
-      }
-
-      if (coverage === "pending") {
-        this.pendingFileChangeAfterToolCallObservations.set(item.id, item);
-        return;
-      }
-    }
-
-    this.emitAfterToolCallObservationNow(item);
+    this.afterToolCall.emit(item);
   }
 
   settlePendingFileChangeAfterToolCallObservations(options?: { finalize?: boolean }): void {
-    for (const [itemId, item] of this.pendingFileChangeAfterToolCallObservations) {
-      const coverage =
-        this.options.resolveNativeFileChangeAfterToolCallCoverage?.(itemId) ?? "pending";
-
-      if (coverage === "native_apply_patch") {
-        this.afterToolCallObservedItemIds.add(itemId);
-        this.pendingFileChangeAfterToolCallObservations.delete(itemId);
-        continue;
-      }
-
-      if (coverage === "intercepted" || options?.finalize === true) {
-        this.pendingFileChangeAfterToolCallObservations.delete(itemId);
-        this.emitAfterToolCallObservationNow(item);
-      }
-    }
-  }
-
-  private emitAfterToolCallObservationNow(item: CodexThreadItem): void {
-    if (!this.shouldEmitAfterToolCallObservation(item)) {
-      return;
-    }
-
-    const name = itemName(item);
-    const status = itemStatus(item);
-
-    if (!name || status === "running") {
-      return;
-    }
-
-    this.afterToolCallObservedItemIds.add(item.id);
-
-    const result = itemToolResult(item).result;
-    const error =
-      this.progress.approvalTimeoutExplanation(item.id, status) ??
-      itemToolError(item, status, this.progress.outputTextByItem);
-
-    const startedAt = resolveStartedAtFromDurationMs(item.durationMs);
-
-    const hookParams = {
-      toolName: name,
-      toolCallId: item.id,
-      runId: this.params.runId,
-      agentId: this.params.agentId,
-      sessionId: this.params.sessionId,
-      sessionKey: this.params.sessionKey,
-      startArgs: itemToolArgs(item) ?? {},
-      ...(result !== undefined ? { result } : {}),
-      ...(error ? { error } : {}),
-      ...(startedAt !== undefined ? { startedAt } : {}),
-    };
-
-    setImmediate(() => {
-      void runAgentHarnessAfterToolCallHook(hookParams);
-    });
+    this.afterToolCall.settlePendingFileChanges(options);
   }
 
   synthesizeMissingToolResults(params: {
@@ -713,24 +632,6 @@ export class CodexToolTranscriptProjection {
     });
   }
 
-  private shouldEmitAfterToolCallObservation(item: CodexThreadItem): boolean {
-    if (
-      !shouldSynthesizeToolProgressForItem(item) ||
-      this.afterToolCallObservedItemIds.has(item.id)
-    ) {
-      return false;
-    }
-    if (!this.options.nativePostToolUseRelayEnabled) {
-      return true;
-    }
-
-    if (item.type === "fileChange") {
-      return true;
-    }
-
-    return !isNativePostToolUseRelayItem(item);
-  }
-
   private createToolCallMessage(params: ToolTranscriptCallInput): AgentMessage {
     const args = normalizeToolTranscriptArguments(params.arguments);
     const attribution = resolveCodexLocalRuntimeAttribution(this.params);
@@ -785,11 +686,4 @@ function formatMissingToolResultError(params: { id: string; name: string }): str
 
 function toolResultStatusText(params: ToolTranscriptResultInput): string {
   return params.isError ? `${params.name} failed` : `${params.name} completed`;
-}
-
-function resolveStartedAtFromDurationMs(durationMs: unknown): number | undefined {
-  if (typeof durationMs !== "number" || !Number.isFinite(durationMs)) {
-    return undefined;
-  }
-  return asDateTimestampMs(Date.now() - Math.max(0, durationMs));
 }
