+++ test/vitest/vitest.extension-codex-app-server-attempt-extra.config.ts
@@ -15,6 +15,7 @@ export function createExtensionCodexAppServerAttemptExtraVitestConfig(
       "extensions/codex/src/app-server/run-attempt.dynamic-tools.test.ts",
       "extensions/codex/src/app-server/run-attempt.hooks.test.ts",
       "extensions/codex/src/app-server/run-attempt.native-hook-relay.test.ts",
+      "extensions/codex/src/app-server/run-attempt.native-hook-relay-filechange.test.ts",
       "extensions/codex/src/app-server/run-attempt.native-hook-relay-retention.test.ts",
       "extensions/codex/src/app-server/run-attempt.notification-burst.test.ts",
       "extensions/codex/src/app-server/run-attempt.reasoning-effort.test.ts",
